feat(bau update script)
Created and populated create_combined_posts_cron.R which reads in all posts, identifies new ones, processes the new ones, and then joins them on to the previous weeks' processed data. The reason for this is that the httr processing section is computationally intensive so re-running this would be inefficient.